finishing touches to provider threshold notifications
