+++ dww_backend/api/views/provider_views.py
@@ -207,45 +207,3 @@ def add_patient_info(request):
     
     print(f'Serializer errors: {serializer.errors}')
     return JsonResponse({'error': 'Invalid JSON'}, status=400) 
-
-def send_weight_change_notification(patient, weight_change, providers): # need to call this function when the drastic weight changes is detected
-    subject = f"Alert: Drastic Weight Change for {patient.first_name} {patient.last_name}"
-
-    message_content = f"""
-    <html>
-      <body style="font-family: Arial, sans-serif; background-color: #f9f9f9; margin: 0; padding: 20px;">
-        <table width="100%" cellpadding="0" cellspacing="0" style="max-width: 600px; margin: 0 auto; background-color: #ffffff; border: 1px solid #ddd; border-radius: 8px; padding: 20px;">
-          <tr>
-            <td style="text-align: center; padding-bottom: 20px;">
-              <h1 style="color: #333; font-size: 24px;">Weight Change Alert</h1>
-              <p style="color: #555; font-size: 16px;">Patient {patient.first_name} {patient.last_name} has experienced a significant weight change.</p>
-              <p style="color: #555; font-size: 16px;">Previous Weight: {weight_change['previous_weight']} kg</p>
-              <p style="color: #555; font-size: 16px;">New Weight: {weight_change['new_weight']} kg</p>
-              <p style="color: #555; font-size: 16px;">Change: {weight_change['change']} kg</p>
-            </td>
-          </tr>
-          <tr>
-            <td style="text-align: center; padding-top: 20px; font-size: 14px; color: #999;">
-              Please review the patient's data and take appropriate action if needed.
-            </td>
-          </tr>
-        </table>
-      </body>
-    </html>
-    """
-
-    provider_emails = [provider.email for provider in providers]
-
-    from_email = settings.EMAIL_HOST_USER
-
-    try:
-        send_mail(
-            subject,
-            '',
-            from_email,
-            provider_emails,
-            html_message=message_content,
-            fail_silently=False,
-        )
-    except Exception as e:
-        raise e
